ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
pydantic==2.5.0
google-generativeai==0.3.2 